FISH-6974 Fix class packaging
